+++ src/classes/queue-events.ts
@@ -1,9 +1,12 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-import { CONNECTION_CLOSED_ERROR_MSG } from 'ioredis/built/utils';
 import { QueueEventsOptions } from '../interfaces';
-import { array2obj, delay, isRedisInstance } from '../utils';
+import {
+  array2obj,
+  delay,
+  isNotConnectionError,
+  isRedisInstance,
+} from '../utils';
 import { StreamReadRaw } from '../interfaces/redis-streams';
+import { DELAY_TIME_5 } from '../utils';
 import { QueueBase } from './queue-base';
 import { RedisClient } from './redis-connection';
 
@@ -266,10 +269,10 @@ export class QueueEvents extends QueueBase {
           }
         }
       } catch (err) {
-        if ((err as Error).message !== CONNECTION_CLOSED_ERROR_MSG) {
+        if (isNotConnectionError(err)) {
           throw err;
         }
-        await delay(5000);
+        await delay(DELAY_TIME_5);
       }
     }
   }